+++ conf/script/src/build_system/cmd/compiler/host/get_info/version/__init__.py
@@ -1,2 +1,2 @@
-from build_system.cmd.compiler.host.get_info.version import compiler
+from build_system.cmd.compiler.host.get_info.version import generic_fetch
 from build_system.cmd.compiler.host.get_info.version.fetch_by_compiler_family import fetch_by_compiler_family
